+++ examples/1-getting-started/src/app/models/RootStore.base.ts
@@ -19,12 +19,26 @@ type Refs = {
   todos: ObservableMap<string, TodoModelType>
 }
 
+
+/**
+* Enums for the names of base graphql actions
+*/
+export enum RootStoreBaseQueries {
+queryTodos="queryTodos",
+queryStringFromServer="queryStringFromServer"
+}
+export enum RootStoreBaseMutations {
+mutateToggleTodo="mutateToggleTodo",
+mutateCreateTodo="mutateCreateTodo",
+mutateReturnBoolean="mutateReturnBoolean"
+}
+
 /**
 * Store, managing, among others, all the objects received through graphQL
 */
 export const RootStoreBase = withTypedRefs<Refs>()(MSTGQLStore
   .named("RootStore")
-  .extend(configureStoreMixin([['Todo', () => TodoModel]], ['Todo']))
+  .extend(configureStoreMixin([['Todo', () => TodoModel]], ['Todo'], "js"))
   .props({
     todos: types.optional(types.map(types.late((): any => TodoModel)), {})
   })
@@ -33,6 +47,9 @@ export const RootStoreBase = withTypedRefs<Refs>()(MSTGQLStore
       return self.query<{ todos: TodoModelType[]}>(`query todos { todos {
         ${typeof resultSelector === "function" ? resultSelector(new TodoModelSelector()).toString() : resultSelector}
       } }`, variables, options)
+    },
+    queryStringFromServer(variables: { string?: string }, options: QueryOptions = {}) {
+      return self.query<{ stringFromServer: string }>(`query stringFromServer($string: String) { stringFromServer(string: $string) }`, variables, options)
     },
     mutateToggleTodo(variables: { id: string }, resultSelector: string | ((qb: TodoModelSelector) => TodoModelSelector) = todoModelPrimitives.toString(), optimisticUpdate?: () => void) {
       return self.mutate<{ toggleTodo: TodoModelType}>(`mutation toggleTodo($id: ID!) { toggleTodo(id: $id) {
@@ -43,5 +60,8 @@ export const RootStoreBase = withTypedRefs<Refs>()(MSTGQLStore
       return self.mutate<{ createTodo: TodoModelType}>(`mutation createTodo($todo: CreateTodoInput!) { createTodo(todo: $todo) {
         ${typeof resultSelector === "function" ? resultSelector(new TodoModelSelector()).toString() : resultSelector}
       } }`, variables, optimisticUpdate)
+    },
+    mutateReturnBoolean(variables: { toReturn: boolean }, optimisticUpdate?: () => void) {
+      return self.mutate<{ returnBoolean: boolean }>(`mutation returnBoolean($toReturn: Boolean!) { returnBoolean(toReturn: $toReturn) }`, variables, optimisticUpdate)
     },
   })))