+++ generator/generate.js
@@ -1,6 +1,8 @@
 const path = require("path")
 const fs = require("fs")
 const graphql = require("graphql")
+const camelcase = require("camelcase")
+const pluralize = require("pluralize")
 
 const exampleAction = `  .actions(self => ({
     // This is an auto-generated example action.
@@ -9,7 +11,7 @@ const exampleAction = `  .actions(self => ({
     }
   }))`
 
-const buildInExcludes = [
+const reservedGraphqlNames = [
   "Mutation",
   "CacheControlScope",
   "Query",
@@ -17,27 +19,42 @@ const buildInExcludes = [
 ]
 
 function generate(
-  types,
+  schema,
   format = "js",
   rootTypes = [],
   excludes = [],
   generationDate = "a long long time ago...",
   modelsOnly = false,
-  noReact = false
+  noReact = false,
+  namingConvention = "js"
 ) {
-  excludes.push(...buildInExcludes)
+  const types = schema.types
+
+  excludes.push(...reservedGraphqlNames)
+
+  // For each type converts 'name' according to namingConvention and copies
+  // original name to the 'origName' field of the type's object
+  transformTypes(types, namingConvention)
 
   const files = [] // [[name, contents]]
   const objectTypes = [] // all known OBJECT types for which MST classes are generated
+  const unionTypes = [] // all known UNION types for which ModelSelector classes are generated
+  const origObjectTypes = [] // all known OBJECT types for which MST classes are generated
   const inputTypes = [] // all known INPUT_OBJECT types for which MST classes are generated
   const knownTypes = [] // all known types (including enums and such) for which MST classes are generated
   const enumTypes = [] // enum types to be imported when using typescript
   const toExport = [] // files to be exported from barrel file
   let currentType = "<none>"
+  let origRootTypes = []
 
   const header = `/* This is a @kibeo/mst-gql generated file, don't modify it manually */
 /* eslint-disable */${format === "ts" ? "\n/* tslint:disable */" : ""}`
   const importPostFix = format === "mjs" ? ".mjs" : ""
+
+  // For a model called TodoModel the TS type would originally be called TodoModelType.
+  // Now we make it just "Todo".  But I have this variable so that in case we want to
+  // make it configurable we can just set this back to "ModelType" or to something else.
+  const modelTypePostfix = "ModelType"
 
   const interfaceAndUnionTypes = resolveInterfaceAndUnionTypes(types)
 
@@ -65,8 +82,10 @@ export const ModelBase = MSTGQLObject
       .filter(type => type.kind !== "SCALAR")
       .forEach(type => {
         knownTypes.push(type.name)
-        if (type.kind === "OBJECT") objectTypes.push(type.name)
-        else if (type.kind === "INPUT_OBJECT") inputTypes.push(type)
+        if (type.kind === "OBJECT") {
+          origObjectTypes.push(type.origName)
+          objectTypes.push(type.name)
+        } else if (type.kind === "INPUT_OBJECT") inputTypes.push(type)
       })
 
     if (!rootTypes.length) {
@@ -78,11 +97,23 @@ export const ModelBase = MSTGQLObject
     }
 
     rootTypes.forEach(type => {
-      if (!objectTypes.includes(type))
+      if (!origObjectTypes.includes(type)) {
+        if (isTypeReservedName(type)) {
+          throw new Error(
+            `Cannot generate ${type}Model, ${type} is a graphql reserved name`
+          )
+        }
         throw new Error(
           `The root type specified: '${type}' is unknown, excluded or not an OBJECT type!`
         )
+      }
     })
+
+    // Keep the orig type names for mixin configuration
+    origRootTypes = [...rootTypes]
+    rootTypes = rootTypes.map(t => transformTypeName(t, namingConvention))
+
+    console.log("rootTypes", JSON.stringify(rootTypes, null, 2))
 
     types
       .filter(type => knownTypes.includes(type.name))
@@ -113,6 +144,10 @@ export const ModelBase = MSTGQLObject
     return type.kind === "NON_NULL" ? type.ofType : type
   }
 
+  function isTypeReservedName(typeName) {
+    return reservedGraphqlNames.includes(typeName)
+  }
+
   function autoDetectRootTypes() {
     return types
       .filter(
@@ -125,12 +160,13 @@ export const ModelBase = MSTGQLObject
               skipNonNull(field.type).name === "ID"
           )
       )
-      .map(t => t.name)
+      .map(t => t.origName)
   }
 
   function handleEnumType(type) {
     const name = type.name
-    toExport.push(name + "Enum")
+    const enumPostfix = !name.toLowerCase().endsWith("enum") ? "Enum" : ""
+    toExport.push(name + enumPostfix)
 
     const tsType =
       format === "ts"
@@ -153,12 +189,12 @@ ${tsType}
 /**
 * ${name}${optPrefix("\n *\n * ", sanitizeComment(type.description))}
 */
-export const ${name}Enum = ${handleEnumTypeCore(type)}
+export const ${name}${enumPostfix}Type = ${handleEnumTypeCore(type)}
 `
     if (format === "ts") {
       enumTypes.push(type.name)
     }
-    generateFile(name + "Enum", contents, true)
+    generateFile(name + enumPostfix, contents, true)
   }
 
   function handleEnumTypeCore(type) {
@@ -184,7 +220,7 @@ export const ${name}Enum = ${handleEnumTypeCore(type)}
       refs
     } = resolveFieldsAndImports(type)
 
-    const { name } = type
+    const { name, origName } = type
     const flowerName = toFirstLower(name)
 
     const entryFile = `${ifTS('import { Instance } from "mobx-state-tree"\n')}\
@@ -192,7 +228,7 @@ import { ${name}ModelBase } from "./${name}Model.base${importPostFix}"
 
 ${
   format === "ts"
-    ? `/* The TypeScript type of an instance of ${name}Model */\nexport interface ${name}ModelType extends Instance<typeof ${name}Model.Type> {}\n`
+    ? `/* The TypeScript type of an instance of ${name}Model */\nexport interface ${name}${modelTypePostfix} extends Instance<typeof ${name}Model.Type> {}\n`
     : ""
 }
 ${
@@ -237,8 +273,8 @@ type Refs = {
 ${refs
   .map(([fieldName, fieldTypeName, isNested]) =>
     isNested
-      ? `  ${fieldName}: IObservableArray<${fieldTypeName}ModelType>;`
-      : `  ${fieldName}: ${fieldTypeName}ModelType;`
+      ? `  ${fieldName}: IObservableArray<${fieldTypeName}${modelTypePostfix}>;`
+      : `  ${fieldName}: ${fieldTypeName}${modelTypePostfix};`
   )
   .join("\n")}
 }\n
@@ -257,7 +293,7 @@ export const ${name}ModelBase = ${
     }ModelBase
   .named('${name}')
   .props({
-    __typename: types.optional(types.literal("${name}"), "${name}"),
+    __typename: types.optional(types.literal("${origName}"), "${origName}"),
 ${modelProperties}
   })
   .views(self => ({
@@ -282,6 +318,7 @@ ${generateFragments(name, primitiveFields, nonPrimitiveFields)}
     const interfaceOrUnionType = interfaceAndUnionTypes.get(type.name)
     const isUnion =
       interfaceOrUnionType && interfaceOrUnionType.kind === "UNION"
+    if (isUnion) unionTypes.push(type.name)
     const fileName = type.name + "ModelSelector"
     const {
       primitiveFields,
@@ -291,14 +328,38 @@ ${generateFragments(name, primitiveFields, nonPrimitiveFields)}
 
     interfaceOrUnionType &&
       interfaceOrUnionType.ofTypes.forEach(t => {
+        /** Base file imports */
         const toBeImported = [`${t.name}ModelSelector`]
         if (isUnion) toBeImported.push(`${toFirstLower(t.name)}ModelPrimitives`)
         addImportToMap(imports, fileName, `${t.name}Model`, ...toBeImported)
+
+        /** Imports from core model */
+        if (isUnion) {
+          // Import <model>ModelType from the core file to be used in the TS union type
+          addImportToMap(
+            imports,
+            fileName,
+            `${t.name}Model`,
+            `${t.name}ModelType`
+          )
+        }
       })
 
     let contents = header + "\n\n"
     contents += 'import { QueryBuilder } from "@kibeo/mst-gql"\n'
     contents += printRelativeImports(imports)
+
+    // Add the correct type for a TS union to the exports
+    if (isUnion) {
+      contents += ifTS(
+        `export type ${
+          interfaceOrUnionType.name
+        }Union = ${interfaceOrUnionType.ofTypes
+          .map(unionModel => `${unionModel.name}ModelType`)
+          .join(" | ")}\n\n`
+      )
+    }
+
     contents += generateFragments(
       type.name,
       primitiveFields,
@@ -373,10 +434,18 @@ ${generateFragments(name, primitiveFields, nonPrimitiveFields)}
           )
         case "ENUM":
           primitiveFields.push(fieldName)
-          const enumType = fieldType.name + "Enum"
+          const enumType =
+            fieldType.name +
+            (!fieldType.name.toLowerCase().endsWith("enum")
+              ? "EnumType"
+              : "Type")
           if (type.kind !== "UNION" && type.kind !== "INTERFACE") {
             // TODO: import again when enums in query builders are supported
-            addImport(enumType, enumType)
+            addImport(
+              fieldType.name +
+                (!fieldType.name.toLowerCase().endsWith("enum") ? "Enum" : ""),
+              enumType
+            )
           }
           return result(enumType)
         case "INTERFACE":
@@ -395,12 +464,11 @@ ${generateFragments(name, primitiveFields, nonPrimitiveFields)}
 
       // this type is not going to be handled by @kibeo/mst-gql, store as frozen
       if (!knownTypes.includes(fieldType.name)) return `types.frozen()`
-
       // import the type
       const modelType = fieldType.name + "Model"
       addImport(modelType, modelType)
       if (format === "ts") {
-        addImport(modelType, `${modelType}Type`)
+        addImport(modelType, `${fieldType.name}${modelTypePostfix}`)
       }
       if (!modelsOnly) {
         addImport(`${modelType}`, `${modelType}Selector`)
@@ -552,13 +620,24 @@ import { MSTGQLStore, configureStoreMixin${
 ${objectTypes
   .map(
     t =>
-      `\nimport { ${t}Model${ifTS(`, ${t}ModelType`)}, ${toFirstLower(
+      `\nimport { ${t}Model${ifTS(`, ${t}${modelTypePostfix}`)}, ${toFirstLower(
         t
       )}ModelPrimitives, ${t}ModelSelector  } from "./${t}Model${importPostFix}"`
   )
   .join("")}
+${unionTypes.map(
+  t =>
+    `\nimport { ${toFirstLower(t)}ModelPrimitives, ${t}ModelSelector ${ifTS(
+      `, ${t}Union`
+    )} } from "./"`
+)}
 ${enumTypes
-  .map(t => `\nimport { ${t} } from "./${t}Enum${importPostFix}"`)
+  .map(
+    t =>
+      `\nimport { ${t} } from "./${t}${
+        !t.toLowerCase().endsWith("enum") ? "Enum" : ""
+      }${importPostFix}"`
+  )
   .join("")}
 ${ifTS(
   inputTypes
@@ -575,12 +654,20 @@ type Refs = {
 ${rootTypes
   .map(
     t =>
-      `  ${t.toLowerCase()}${
-        t.charAt(t.length - 1) !== "s" ? "s" : ""
-      }: ObservableMap<string, ${t}ModelType>`
+      `  ${transformRootName(
+        t,
+        namingConvention
+      )}: ObservableMap<string, ${t}${modelTypePostfix}>`
   )
   .join(",\n")}
 }\n\n`)}\
+${ifTS(`
+/**
+* Enums for the names of base graphql actions
+*/`)}
+${ifTS(generateGraphQLActionsEnum("Query", "Queries", "query"))}
+${ifTS(generateGraphQLActionsEnum("Mutation", "Mutations", "mutate"))}
+
 /**
 * Store, managing, among others, all the objects received through graphQL
 */
@@ -588,16 +675,19 @@ export const RootStoreBase = ${ifTS("withTypedRefs<Refs>()(")}${
       modelsOnly ? "types.model()" : "MSTGQLStore"
     }
   .named("RootStore")
-  .extend(configureStoreMixin([${objectTypes
-    .map(s => `['${s}', () => ${s}Model]`)
-    .join(", ")}], [${rootTypes.map(s => `'${s}'`).join(", ")}]))
+  .extend(configureStoreMixin([${origObjectTypes
+    .map(s => `['${s}', () => ${transformTypeName(s, namingConvention)}Model]`)
+    .join(", ")}], [${origRootTypes.map(s => `'${s}'`).join(", ")}]${
+      namingConvention == "asis" ? "" : `, "${namingConvention}"`
+    }))
   .props({
 ${rootTypes
   .map(
     t =>
-      `    ${t.toLowerCase()}${
-        t.charAt(t.length - 1) !== "s" ? "s" : ""
-      }: types.optional(types.map(types.late(()${ifTS(
+      `    ${transformRootName(
+        t,
+        namingConvention
+      )}: types.optional(types.map(types.late(()${ifTS(
         ": any"
       )} => ${t}Model)), {})`
   ) // optional should not be needed here..
@@ -610,11 +700,64 @@ ${rootTypes
     generateFile("RootStore.base", modelFile, true)
   }
 
+  /**
+   * Returns if this field should be skipped in generation. Can happen if:
+   * 1) The field is in the excludes
+   * 2) The field has a return type that is not supported
+   * @param {*} field from an array of queries or mutations
+   */
+  function shouldSkipField(field) {
+    let { name, origName, args, type, description } = field
+
+    if (type.kind === "NON_NULL") type = type.ofType
+    const returnsList = type.kind === "LIST"
+    let returnType = returnsList ? type.ofType : type
+    if (returnType.kind === "NON_NULL") returnType = returnType.ofType
+
+    if (returnType.kind === "OBJECT" && excludes.includes(returnType.name))
+      return true
+    // TODO: probably we will need to support input object types soon
+    return false
+  }
+
+  /**
+   * A func to generate enums that are the names of the graphql actions in the RootStore.base
+   * Like:
+   * export enum RootStoreBaseQueries {
+   *    queryMessages="queryMessages",
+   *    queryMessage="queryMessage",
+   *    queryMe="queryMe"
+   * }
+   *
+   *
+   * @param {*} gqlType Query | Mutation
+   * @param {*} gqlPrefix query | mutation
+   */
+  function generateGraphQLActionsEnum(gqlType, gqlPlural, methodPrefix) {
+    const queries = findObjectByName(gqlType)
+    if (!queries) return ""
+
+    const enumContent = queries.fields
+      .map(field => {
+        const { name } = field
+        if (shouldSkipField(field)) return ""
+        const queryName = `${methodPrefix}${toFirstUpper(name)}`
+        return `${queryName}="${queryName}"`
+      })
+      // Filter out empty strings for skipped fields
+      .filter(n => n)
+      .join(",\n")
+    if (enumContent === "") return
+    return `export enum RootStoreBase${gqlPlural} {
+${enumContent}
+}`
+  }
+
   function generateQueries() {
     if (modelsOnly) return ""
     return (
       generateQueryHelper(
-        findObjectByName("Query"),
+        findObjectByName(schema.queryType ? schema.queryType.name : "Query"),
         "query",
         "query",
         format === "ts"
@@ -623,7 +766,9 @@ ${rootTypes
         ", options, !!clean"
       ) +
       generateQueryHelper(
-        findObjectByName("Mutation"),
+        findObjectByName(
+          schema.mutationType ? schema.mutationType.name : "Mutation"
+        ),
         "mutation",
         "mutate",
         format === "ts"
@@ -632,13 +777,17 @@ ${rootTypes
         ", optimisticUpdate"
       ) +
       generateQueryHelper(
-        findObjectByName("Subscription"),
+        findObjectByName(
+          schema.subscriptionType
+            ? schema.subscriptionType.name
+            : "Subscription"
+        ),
         "subscription",
         "subscribe",
         format === "ts"
-          ? `, onData?: (item: any) => void` /* TODO: fix the any */
-          : `, onData`,
-        ", onData"
+          ? `, onData?: (item: any) => void, onError?: (error: Error) => void` /* TODO: fix the any */
+          : `, onData, onError`,
+        ", onData, onError"
       )
     )
   }
@@ -651,34 +800,32 @@ ${rootTypes
     extraActualArgs = ""
   ) {
     if (!query) return ""
+
     return query.fields
       .map(field => {
-        let { name, args, type, description } = field
+        if (shouldSkipField(field)) return ""
+
+        let { name, origName, args, type, description } = field
+
+        const isScalar =
+          type.kind === "SCALAR" ||
+          (type.ofType && type.ofType.kind === "SCALAR")
 
         if (type.kind === "NON_NULL") type = type.ofType
         const returnsList = type.kind === "LIST"
         let returnType = returnsList ? type.ofType : type
         if (returnType.kind === "NON_NULL") returnType = returnType.ofType
 
-        if (returnType.kind === "OBJECT" && excludes.includes(returnType.name))
-          return ""
-        // TODO: probably we will need to support input object types soon
-        // if (returnType.kind !== "OBJECT") {
-        //   console.warn(
-        //     `Skipping generation of query '${name}', its return type is not yet understood. PR is welcome`
-        //   )
-        //   // log(returnType)
-        //   return "" // TODO: for now, we only generate queries for those queries that return objects
-        // }
-
         const tsType =
           format !== "ts"
             ? ""
             : `<{ ${name}: ${
-                returnType.kind !== "OBJECT"
-                  ? printTsPrimitiveType(returnType.name.toLowerCase())
-                  : `${returnType.name}ModelType`
-              }${returnsList ? "[]" : ""}}>`
+                isScalar
+                  ? `${printTsPrimitiveType(type.name)} `
+                  : `${returnType.name}${
+                      returnType.kind === "UNION" ? "Union" : modelTypePostfix
+                    }${returnsList ? "[]" : ""}`
+              }}>`
 
         const formalArgs =
           args.length === 0
@@ -692,7 +839,7 @@ ${rootTypes
           args.length === 0
             ? ""
             : "(" +
-              args.map(arg => `${arg.name}: \$${arg.name}`).join(", ") +
+              args.map(arg => `${arg.origName}: \$${arg.name}`).join(", ") +
               ")"
 
         const tsVariablesType =
@@ -743,7 +890,7 @@ ${optPrefix("\n    // ", sanitizeComment(description))}
       case "OBJECT":
       case "INPUT_OBJECT":
       case "SCALAR":
-        return type.name
+        return type.origName ? type.origName : type.name
       default:
         throw new Error(
           "Not implemented printGraphQLType yet, PR welcome for " +
@@ -752,7 +899,12 @@ ${optPrefix("\n    // ", sanitizeComment(description))}
     }
   }
 
-  function printTsType(field, name, canBeUndefined = true) {
+  function printTsType(
+    field,
+    name,
+    canBeUndefined = true,
+    fromUndefineableList = false
+  ) {
     let typeValue
     let type
 
@@ -765,9 +917,9 @@ ${optPrefix("\n    // ", sanitizeComment(description))}
 
     switch (type.kind) {
       case "NON_NULL":
-        return printTsType(type.ofType, name, false)
+        return printTsType(type.ofType, name, false, fromUndefineableList)
       case "LIST":
-        return `${printTsType(type.ofType, name, true)}[]`
+        return `${printTsType(type.ofType, name, true, canBeUndefined)}[]`
       case "OBJECT":
       case "INPUT_OBJECT":
       case "ENUM":
@@ -784,9 +936,9 @@ ${optPrefix("\n    // ", sanitizeComment(description))}
         typeValue = "any"
     }
 
-    return `${name}${canBeUndefined ? "?" : ""}: ${typeValue} ${
-      canBeUndefined ? "| null" : ""
-    }`
+    return `${name}${
+      canBeUndefined || fromUndefineableList ? "?" : ""
+    }: ${typeValue} ${canBeUndefined ? "| null" : ""}`
   }
 
   function printTsPrimitiveType(primitiveType) {
@@ -802,7 +954,7 @@ ${optPrefix("\n    // ", sanitizeComment(description))}
   }
 
   function findObjectByName(name) {
-    return types.find(type => type.name === name && type.kind === "OBJECT")
+    return types.find(type => type.origName === name && type.kind === "OBJECT")
   }
 
   function generateReactUtils() {
@@ -814,9 +966,7 @@ import { createStoreContext, createUseQueryHook } from "@kibeo/mst-gql"
 import * as React from "react"
 ${
   format === "ts"
-    ? `import { RootStore${ifTS(
-        ", RootStoreType"
-      )} } from "./RootStore${importPostFix}"`
+    ? `import { RootStoreType } from "./RootStore${importPostFix}"`
     : ""
 }
 
@@ -1051,7 +1201,8 @@ function scaffold(
     format: "ts",
     roots: [],
     excludes: [],
-    modelsOnly: false
+    modelsOnly: false,
+    namingConvention: "js"
   }
 ) {
   const schema = graphql.buildSchema(definition)
@@ -1059,13 +1210,138 @@ function scaffold(
   if (!res.data)
     throw new Error("graphql parse error:\n\n" + JSON.stringify(res, null, 2))
   return generate(
-    res.data.__schema.types,
+    res.data.__schema,
     options.format || "ts",
     options.roots || [],
     options.excludes || [],
     "<during unit test run>",
-    options.modelsOnly || false
+    options.modelsOnly || false,
+    options.noReact || false,
+    options.namingConvention || "js"
   )
+}
+
+/**
+ * Transforms {@code text} according to namingConvention.
+ * If namingConvention is {@code null} or {@code asis} the same text will be
+ * returned. If namingConvention is {@code js} text will be PascalCased
+ * @param text to be transformed
+ * @param namingConvention naming convention to be used for transformation
+ * @returns {string|*}
+ */
+function transformTypeName(text, namingConvention) {
+  if (!text) {
+    return text
+  }
+  if (namingConvention === "js") {
+    return camelcase(text, { pascalCase: true })
+  }
+  return text
+}
+
+/**
+ * Transforms {@code text} according to namingConvention.
+ * If namingConvention is {@code null} or {@code asis} the same text will be
+ * returned. If namingConvention is {@code js} text will be camelCased
+ * @param text to be transformed
+ * @param namingConvention naming convention to be used for transformation
+ * @returns {string|*}
+ */
+function transformName(text, namingConvention) {
+  if (!text) {
+    return text
+  }
+  if (namingConvention === "js") {
+    return camelcase(text)
+  }
+  return text
+}
+
+/**
+ * Transform {@code text} according to namingConvention.
+ * If namingConvention is {@code null} or {@code asis} the same text will be
+ * returned. If namingConvention is {@code js} text will be camelCased and
+ * pluralized
+ * @param text to be transformed
+ * @param namingConvention naming convention to be used for transformation
+ * @returns {string|*}
+ */
+function transformRootName(text, namingConvention) {
+  if (!text) {
+    return text
+  }
+  if (namingConvention === "js") {
+    // Pluralize only last word (pluralize may fail with words that are
+    // not valid English words as is the case with LongCamelCaseTypeNames)
+    const newName = transformName(text, namingConvention)
+    const parts = newName.split(/(?=[A-Z])/)
+    parts[parts.length - 1] = pluralize(parts[parts.length - 1])
+    return parts.join("")
+  }
+  return text.toLowerCase() + "s"
+}
+
+/**
+ * Converts types names in the graphql schema {@code types} according to
+ * the {@code namingConvention}. If namingConvention is {@code null} or
+ * {@code asis} type names won't be transformed.
+ * If namingConvention is {@code js} type named will be PascalCased and the
+ * original type names will be stored in the {@code origName} field.
+ * @param types graphql schema types
+ * @param namingConvention
+ */
+function transformTypes(types, namingConvention) {
+  //console.log(JSON.stringify(types, null, 2));
+  types
+    .filter(type => !type.name.startsWith("__"))
+    .filter(type => type.kind !== "SCALAR")
+    .forEach(type => transformType(type, namingConvention))
+  //console.log(JSON.stringify(types, null, 2));
+}
+
+/**
+ * Transforms names of a graphql {@code type} object according to
+ * {@code namingConvention}. Name of types of kind OBJECT, INPUT_OBJECT,
+ * ENUM, LIST and NON_NULL and changed according to {@code namingConvention}
+ * recursively. If namingConvention is {@code js} type named will be
+ * PascalCased and the original type names will be stored in the
+ * {@code origName} field.
+ * @param type a graphql type definition object
+ * @param namingConvention
+ */
+function transformType(type, namingConvention) {
+  if (
+    type.kind === "OBJECT" ||
+    type.kind === "INPUT_OBJECT" ||
+    type.kind === "ENUM" ||
+    type.kind === "LIST" ||
+    type.kind === "NON_NULL"
+  ) {
+    type.origName = type.name
+    type.name = transformTypeName(type.name, namingConvention)
+
+    // process type names in fields, inputFields and ofType
+    if (type.fields) {
+      type.fields.forEach(f => {
+        // process own type
+        transformType(f.type, namingConvention)
+        // process types of args
+        if (f.args) {
+          f.args.forEach(arg => {
+            arg.origName = arg.name
+            arg.name = transformName(arg.name, namingConvention)
+            transformType(arg.type, namingConvention)
+          })
+        }
+      })
+    }
+    if (type.inputFields) {
+      type.inputFields.forEach(f => transformType(f.type, namingConvention))
+    }
+    if (type.ofType) {
+      transformType(type.ofType, namingConvention)
+    }
+  }
 }
 
 function logUnexpectedFiles(outDir, files) {