+++ src/MSTGQLStore.ts
@@ -1,11 +1,13 @@
-import { SubscriptionClient } from "subscriptions-transport-ws"
-import { types, getEnv, recordPatches, IAnyModelType } from "mobx-state-tree"
+import camelcase from "camelcase"
 import { DocumentNode } from "graphql"
+import { getEnv, IAnyModelType, recordPatches, types } from "mobx-state-tree"
+import pluralize from "pluralize"
+import { SubscriptionClient } from "subscriptions-transport-ws"
+import { deflateHelper } from "./deflateHelper"
 
 import { mergeHelper } from "./mergeHelper"
+import { Query, QueryOptions } from "./Query"
 import { getFirstValue } from "./utils"
-import { QueryOptions, Query } from "./Query"
-import { deflateHelper } from "./deflateHelper"
 
 export interface RequestHandler<T = any> {
   request(query: string, variables: any): Promise<T>
@@ -122,7 +124,10 @@ export const MSTGQLStore = types
     function subscribe<T = any>(
       query: string,
       variables?: any,
-      onData?: (item: T) => void
+      onData?: (item: T) => void,
+      onError: (error: Error) => void = error => {
+        throw error
+      }
     ): () => void {
       if (!gqlWsClient) throw new Error("No WS client available")
       const sub = gqlWsClient
@@ -132,7 +137,9 @@ export const MSTGQLStore = types
         })
         .subscribe({
           next(data) {
-            if (data.errors) throw new Error(JSON.stringify(data.errors))
+            if (data.errors) {
+              return onError(new Error(JSON.stringify(data.errors)))
+            }
             ;(self as any).__runInStoreContext(() => {
               const res = (self as any).merge(getFirstValue(data.data))
               if (onData) onData(res)
@@ -170,7 +177,8 @@ export const MSTGQLStore = types
 
 export function configureStoreMixin(
   knownTypes: [string, () => IAnyModelType][],
-  rootTypes: string[]
+  rootTypes: string[],
+  namingConvention?: string
 ) {
   const kt = new Map()
   const rt = new Set(rootTypes)
@@ -197,6 +205,17 @@ export function configureStoreMixin(
       },
       getTypeDef(typename: string): IAnyModelType {
         return kt.get(typename)!
+      },
+      getCollectionName(typename: string): string {
+        if (namingConvention == "js") {
+          // Pluralize only last word (pluralize may fail with words that are
+          // not valid English words as is the case with LongCamelCaseTypeNames)
+          const newName = camelcase(typename)
+          const parts = newName.split(/(?=[A-Z])/)
+          parts[parts.length - 1] = pluralize(parts[parts.length - 1])
+          return parts.join("")
+        }
+        return typename.toLowerCase() + "s"
       }
     }
   })